+++ crates/cli/src/commands/uninstall.rs
@@ -39,18 +39,6 @@ fn unpin_version(session: &ProtoSession, args: &UninstallArgs) -> Result<(), Pro
                 {
                     doc.as_table_mut().remove(args.context.as_str());
                 }
-
-                // if let Some(versions) = &mut config.versions {
-                //     let remove = if let Some(version) = versions.get(&args.id) {
-                //         args.spec.is_none() || args.spec.as_ref().is_some_and(|spec| spec == version)
-                //     } else {
-                //         false
-                //     };
-
-                //     if remove {
-                //         versions.remove(&args.id);
-                //     }
-                // }
             })?;
         }
     }
@@ -81,12 +69,11 @@ async fn track_uninstall(tool: &Tool, all: bool) -> Result<(), ProtoCliError> {
 #[instrument(skip(session))]
 async fn uninstall_all(session: ProtoSession, args: UninstallArgs) -> AppResult {
     let mut tool = session.load_tool(&args.context).await?;
-    let inventory_dir = tool.get_inventory_dir();
     let version_count = tool.inventory.manifest.installed_versions.len();
     let skip_prompts = session.should_skip_prompts();
     let mut confirmed = false;
 
-    if !inventory_dir.exists() {
+    if !tool.get_inventory_dir().exists() {
         session.console.render(element! {
             Notice(variant: Variant::Caution) {
                 StyledText(
@@ -110,7 +97,7 @@ async fn uninstall_all(session: ProtoSession, args: UninstallArgs) -> AppResult 
                         "Uninstall all {} versions of {} at <path>{}</path>?",
                         version_count,
                         tool.get_name(),
-                        inventory_dir.display()
+                        tool.get_inventory_dir().display()
                     ),
                     on_confirm: &mut confirmed,
                 )
@@ -137,7 +124,7 @@ async fn uninstall_all(session: ProtoSession, args: UninstallArgs) -> AppResult 
     }
 
     // Delete inventory
-    fs::remove_dir_all(inventory_dir)?;
+    fs::remove_dir_all(tool.get_inventory_dir())?;
     fs::remove_dir_all(tool.get_temp_dir())?;
 
     // Remove from lockfile