+++ crates/core/src/layout/shim_registry.rs
@@ -11,14 +11,16 @@ pub struct Shim {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub after_args: Vec<String>,
 
-    #[serde(skip_serializing_if = "Option::is_none")]
+    // TODO rename to alt_exe in next version
+    #[serde(skip_serializing_if = "Option::is_none", alias = "alt_bin")]
     pub alt_bin: Option<bool>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub parent: Option<String>,
 
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub before_args: Vec<String>,
+
+    // TODO rename to context in next version
+    #[serde(skip_serializing_if = "Option::is_none", alias = "parent")]
+    pub parent: Option<String>,
 
     #[serde(skip_serializing_if = "FxHashMap::is_empty")]
     pub env_vars: FxHashMap<String, String>,