+++ package/src/api-types.ts
@@ -44,7 +44,16 @@ export interface ExecCommandInput {
 	working_dir?: VirtualPath | null;
 	/** Override the current working directory. */
 	cwd?: VirtualPath | null;
-	/** Environment variables to pass to the command. */
+	/**
+	 * Environment variables to pass to the command. Variables
+	 * can customize behavior by appending one of the following
+	 * characters to the name:
+	 *
+	 * `?` - Will only set variable if it doesn't exist
+	 * in the current environment.
+	 * `!` - Will remove the variable from being inherited
+	 * by the child process.
+	 */
 	env?: Record<string, string>;
 	/** Mark the command as executable before executing. */
 	setExecutable?: boolean;
@@ -124,6 +133,11 @@ export interface ToolInventoryOptions {
 	 * This is an advanced feature and should only be used when absolutely necessary.
 	 */
 	overrideDir?: VirtualPath | null;
+	/**
+	 * When the inventory is backend managed, scope the inventory directory name
+	 * with the backend as a prefix.
+	 */
+	scopedBackendDir?: boolean;
 	/** Suffix to append to all versions when labeling directories. */
 	versionSuffix?: string | null;
 }
@@ -139,10 +153,12 @@ export interface ToolLockOptions {
 	noRecord?: boolean;
 }
 
+export type Id = string;
+
 /** Input passed to the `register_tool` function. */
 export interface RegisterToolInput {
 	/** ID of the tool, as it was configured. */
-	id: string;
+	id: Id;
 }
 
 /** Supported strategies for installing a tool. */
@@ -205,7 +221,7 @@ export interface RegisterBackendInput {
 	/** Current tool context. */
 	context: PluginUnresolvedContext;
 	/** ID of the tool, as it was configured. */
-	id: string;
+	id: Id;
 }
 
 /** Source code is contained in an archive. */
@@ -233,7 +249,7 @@ export type SourceLocation = ArchiveSource | GitSource;
 /** Output returned by the `register_backend` function. */
 export interface RegisterBackendOutput {
 	/** Unique identifier for this backend. Will be used as the folder name. */
-	backendId: string;
+	backendId: Id;
 	/**
 	 * List of executables, relative from the backend directory,
 	 * that will be executed in the context of proto.
@@ -241,6 +257,12 @@ export interface RegisterBackendOutput {
 	exes?: string[];
 	/** Location in which to acquire source files for the backend. */
 	source?: SourceLocation | null;
+}
+
+/** Output returned from the `define_backend_config` function. */
+export interface DefineBackendConfigOutput {
+	/** Schema shape of the backend's configuration. */
+	schema: unknown;
 }
 
 /** Output returned by the `detect_version_files` function. */
@@ -276,6 +298,8 @@ export interface ParseVersionFileOutput {
 export interface NativeInstallInput {
 	/** Current tool context. */
 	context: PluginContext;
+	/** Whether to force install or not. */
+	force: boolean;
 	/** Virtual directory to install to. */
 	installDir: VirtualPath;
 }
@@ -593,7 +617,7 @@ export type BuildInstruction = {
 		/** The Git source location for the builder. */
 		git: GitSource;
 		/** Unique identifier for this builder. */
-		id: string;
+		id: Id;
 	};
 	type: 'install-builder';
 } | {